+++ igarden/admin.py
@@ -1,6 +1,5 @@
 from django.contrib import admin
-from .models import Flower, UserFlowersList
+from .models import Flower
 
 # Register your models here.
 admin.site.register(Flower)
-admin.site.register(UserFlowersList)
