+++ igarden/templates/igarden/detail.html
@@ -23,7 +23,7 @@
                         {% endfor %}
                     </div>
                 </div>
-                <a href="{% url 'igarden-create-list' %}">Add new list</a>
+                <a href="{% url 'lists-create-list' %}">Add new list</a>
             </div>
             <div class="col-8">
                 <article class="flower-descr">
