+++ lists/templates/lists.html
@@ -7,7 +7,7 @@
             <div class="card-body">
                 <h4 class="card-title">{{ user.username }}</h4>
                 <a href="{% url 'profile' %}" class="btn btn-primary">See Profile</a>
-                <a href="{% url 'igarden-create-list' %}" class="btn btn-success mt-2">Add new list</a>
+                <a href="{% url 'lists-create-list' %}" class="btn btn-success mt-2">Add new list</a>
             </div>
         </div>
         <div class="col-9">
